Complete code refactor, i18n sweep, test restructure, and future functionality enhancements
- Removed orphaned/dead files: app_features.dart, learning_plan_dashboard.dart,
  analytics_dashboard.dart, subject_management_screen.dart
- Created shared llmServiceProvider in lib/core/providers/llm_providers.dart
- Converted MentorScreen, TutorScreen, QuickGuideScreen to ConsumerStatefulWidget
  so AI screens read user-configured API key instead of silent mock fallback
- Refactored MasteryState to immutable model, extracted MasteryCalculationService
- Moved subject_model.dart to data/models, extracted TokenPricingConfig
- Fixed engagement_scheduler stubs, added Test Connection button to API config,
  added voice/image capture stubs to tutor and upload screens
- Internationalized 60+ new strings in dashboard, upload, planner, llm_task_manager
  screens with full EN/ES ARB translations
- Consolidated test files to proper feature directories, removed dead tests,
  fixed integration test imports
- Hardcoded color sweep: replaced Colors.* with colorScheme tokens across
  dashboard, lessons, practice, sessions, subjects screens
- Replaced all debugPrint with Logger, consolidated AnswerValidationService

+++ lib/core/utils/responsive.dart
@@ -69,7 +69,7 @@ class ResponsiveUtils {
     }
   }
 
-  static double gridCrossAxisCount(BuildContext context) {
+  static int gridCrossAxisCount(BuildContext context) {
     final bp = breakpointOf(context);
     switch (bp) {
       case ScreenBreakpoint.xs:
